refactor(test): move database tests to tests/data/
