+++ tests/test_regex_identifier.py
@@ -32,38 +32,6 @@ def _assert_match_in_items(name, res):
     assert any(name in i["Regex Pattern"]["Name"] for i in res)
 
 
-@pytest.mark.skip(
-    reason="Not all regex have tests now, check https://github.com/bee-san/pyWhat/pull/146#issuecomment-927087231 for info."
-)
-def test_if_all_tests_exist():
-    with open("tests/test_regex_identifier.py", "r", encoding="utf-8") as file:
-        tests = file.read()
-
-    for regex in database:
-        assert (
-            regex["Name"] in tests
-        ), "No test for this regex found in 'test_regex_identifier.py'. Note that a test needs to assert the whole name."
-
-
-def test_regex_format():
-    for regex in database:
-        assert re.findall(
-            r"^(?:\(\?i\))?\^\(.*\)\$$", regex["Regex"]
-        ), r"Please use ^(regex)$ regex format. If there is '\n' character, you have to escape it. If there is '(?i)', it is allowed and should be before the '^'."
-
-        assert (
-            re.findall(r"\^\||\|\^|\$\|\^|\$\||\|\$", regex["Regex"]) == []
-        ), "Remove in-between boundaries. For example, '^|$' should only be '|'."
-
-
-def test_sorted_by_rarity():
-    rarity_num = [regex["Rarity"] for regex in database]
-
-    assert rarity_num == sorted(
-        rarity_num, reverse=True
-    ), "Regexes should be sorted by rarity in 'regex.json'. Regexes with rarity '1' are at the top of the file and '0' is at the bottom."
-
-
 def regex_valid_match(name: str, match: str) -> bool:
     return any(name in matched["Regex Pattern"]["Name"] for matched in r.check([match]))
